+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/ReferenceGenomeImpl.java
@@ -87,4 +87,14 @@ public class ReferenceGenomeImpl implements ReferenceGenome {
     return getId() != UNSAVED_ID;
   }
 
+  @Override
+  public String getDeleteType() {
+    return "Reference Genome";
+  }
+
+  @Override
+  public String getDeleteDescription() {
+    return getAlias();
+  }
+
 }