+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/ReferenceGenomeDto.java
@@ -1,24 +1,33 @@
 package uk.ac.bbsrc.tgac.miso.dto;
 
 public class ReferenceGenomeDto {
-  private String alias;
 
   private Long id;
-
-  public String getAlias() {
-    return alias;
-  }
+  private String alias;
+  private String defaultScientificName;
 
   public Long getId() {
     return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getAlias() {
+    return alias;
   }
 
   public void setAlias(String alias) {
     this.alias = alias;
   }
 
-  public void setId(Long id) {
-    this.id = id;
+  public String getDefaultScientificName() {
+    return defaultScientificName;
+  }
+
+  public void setDefaultScientificName(String defaultScientificName) {
+    this.defaultScientificName = defaultScientificName;
   }
 
 }